rename components to better describe what they are (#673)
