+++ src/lib/components/table-editable/EditableTable.svelte
@@ -5,7 +5,7 @@
    * Its goal it so utilize all of the other container components
    * and provide the interactions needed to do things with the table.
    */
-  import PinnableTable from "./PinnableTable.svelte";
+  import ColumnGroup from "./ColumnGroup.svelte";
   import { createEventDispatcher } from "svelte";
   import { togglePin } from "$lib/components/table-editable/pinnableUtils";
   import type { ColumnConfig } from "$lib/components/table-editable/ColumnConfig";
@@ -24,7 +24,7 @@
 </script>
 
 <div class="flex relative">
-  <PinnableTable
+  <ColumnGroup
     on:pin={handlePin}
     on:change={(evt) => dispatch("change", evt.detail)}
     on:delete={(evt) => dispatch("delete", evt.detail)}
@@ -37,7 +37,7 @@
     <div
       class="sticky right-0 z-20 bg-white border border-l-4 border-t-0 border-b-0 border-r-0 border-gray-300"
     >
-      <PinnableTable
+      <ColumnGroup
         on:pin={handlePin}
         on:change={(evt) => dispatch("change", evt.detail)}
         on:delete={(evt) => dispatch("delete", evt.detail)}
